refactor: move DisjointMutPtr to p3-util
